Steam Cloud Fix
On steam, the game uses Steam Auto-Cloud to synchronize game progress.
The configuration of what steam uploads is not stored locally, so we can't
modify that. With a bit of experimentation we can however determine that
(at the time of writing) it is configured to upload all files in the
profile that end in .owsave. We can take advantage of that and just rename
our file from .json to .owsave. (The actual game save files are is just
json files with other extensions as well, by the way.)

To verify this works, the steam cloud contents can be viewed at
https://store.steampowered.com/account/remotestorageapp/?appid=753640